feat: Create the teacher form
Create the teacher creation form. Create labels, textboxes and masked text boxes to create such a form. Create the Teacher Services and Student Services cs where one would be able to access the load the student / teacher according to the id card number as well as checking if that exists.